requirements.txt
anthropic>=0.25.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
graphql-core>=3.2.0
gql>=3.4.0
aiohttp>=3.8.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0
rich>=13.0.0
typer>=0.9.0
SQLAlchemy>=2.0.0
alembic>=1.12.0
email-validator>=2.0.0
# mem0 integration for conversation memory
mem0ai>=0.0.1
requests>=2.31.0
strands-agents>=1.12.0
